requirements.txt
langchain
langchain-community
llama-cpp-python
streamlit

